face-recognition/requirements.txt
boto3
dlib
numpy
face_recognition
opencv-python
ffmpeg
pandas
python-dotenv